test(`getaddressdeltas`): extract module to separate file
Move address_deltas module from inline in state_service.rs to:
  integration-tests/tests/zebra/get/address_deltas.rs

Module path resolution:
- state_service.rs declares: mod zebra { mod get { mod address_deltas; } }
- Rust looks for: tests/zebra/get/address_deltas.rs (relative to test file)

Benefits:
- Cleaner state_service.rs (removed ~250 lines)
- Dedicated file for address_deltas tests
- Easier to find and modify related tests
- Follows project's preference against mod.rs pattern
